requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-huggingface
chromadb
sentence-transformers
groq
requests
python-dotenv